Do not query health of PLDM fans
The PLDM driver does not have the capability to query health status of
fans in MEX drawers. This change returns early if the serviceName
indicates it came from PLDM.

Signed-off-by: Mike Capps <mikepcapps@gmail.com>

Do not query health of PLDM fans

The PLDM driver does not have the capability to query health status of
fans in MEX drawers. This change returns early if the serviceName
indicates it came from PLDM.

Signed-off-by: Mike Capps <mikepcapps@gmail.com>