Fix PCIe Slot Count Check
When we have more than one chassis in the system, the code
would wrongly compare the total number of slots across all the chassis
to the number of slots in the PATCH request's payload (which is for a
specific chassis). This would cause the PATCH to fail without any
error.

This commit fixes the count check by properly counting the slots that
don't belong in the chassis of interest.

Tested:
Tested this on a system with Mex chassis and a CEC chassis.
Verified that the PATCH on LocationIndicatorActive now works fine
even when we have two chassis trees on D-Bus.

Signed-off-by: Santosh Puranik <santosh.puranik@in.ibm.com>