Remove old Power#/Thermal# from MetricDefinitions (#358)
IBM has moved to using new Power and Thermal schemas and the older
schemas (i.e. Power#/Thermal#) are outdated and not supported.
Additionally, this speeds up the time for a GET call on each
MetricDefinition. The savings are around 40%, moving from about 4.5
seconds to about 2.5 seconds per metric.

Testing:
1) Equal counts in MetricDefinitions and Chassis/chassis/sensors
2) Passed Redfish validator
3) MetricDefinitions remain the samw except removal of references
to the older schemas.