Restrict host console to ConfigureManager (#407)
Currently ReadOnly has access to the host console.
Restrict it to Admin or above.

Part of
[SW553027](https://w3.rchland.ibm.com/projects/bestquest/?defect=SW553027)

Currently, ReadOnly has access to the host console.
Restrict it to Admin or above.

This appears to add a half-second delay in starting the websocket (I
assume this is for the dbus call). It is just noticeable but hard to
measure (e.g. under a second).. Or I am going crazy.

I followed code in
https://github.com/ibm-openbmc/bmcweb/blob/8044381620b228f99b707ea2b444c7675e714f26/include/nbd_proxy.hpp#L417

Tested: For a ReadOnly user I see
```
websocket console0/ closed.
            code: 1000
            reason: Unathourized access
```
in the browser development tools console.

For Admin and Service the console works as normal as does the "Open in
new tab".
In the browser development tools console I see websocket console0/
opened and what I would expect flowing when booting the system.

Signed-off-by: Gunnar Mills <gmills@us.ibm.com>