Bump our connection timeout to 26 mins
Before this was 12 mins but have seen real examples of slow networks
hitting this 12 min timeout. Bump this timeout to 26 mins which should
support an upload speed of as low as 0.641 Mbit/s.

This will be all connections, not just code update.

This is definitely a bandage, long term:
1. code update connection timeout should not be the same as other
connection timeouts
2. We should install a lower code update timeout (e.g. 30 seconds) that
gets reset every time we get a packet
3. We should look at our image size, and make sure we can't make that
smaller

Signed-off-by: Gunnar Mills <gmills@us.ibm.com>