Review fixes for PowerSupplyMetrics
Fix typos.

No need to pass chassisID and powerSupplyID down to getValues(). Change
it to pass down the validated path retrieved via getValidPowerSupplyID().

GetSubtree -> GetObject. The GetSubTree mapper call is likely to be an
"expensive" operation. I already have associations, so use GetObject
instead.

Create a separate getServiceAndPath function. Split off the code for
getting the service name and the object paths into a separate helper
function. This allows to use it for both paths in a common way, avoids
excessive indentation, and makes for fewer lines per function.

Not noted in the review, but I have defined averageInterface and
maximumInterface constants for the strings, so I should use those.

Change-Id: I2611dec59f55f116ac10d174308a506fcfc8b296
Signed-off-by: Brandon Wyman <bjwyman@gmail.com>