Update schema to enable panel CM
The commit makes required change to enable base and LCD
panel CM on Everest systems.
To achieve that whenever the FRU is set for removal,
CCIN property needs to be cleared as some sensors are
dependent of panel presence and listening for CCIN property
of the FRU.
Also post the FRU is replaced, vpd-manager is trigerred
to collect VPD of the FRU and restore its CCIN value and
present property.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>