Follow upstream schema handling
This updates schema handling as to upstream including

- Additional schemas
   AggregationService
   AggregationSourceCollection
   FabricAdapter
   FabricAdapterCollection

- Aggregation: Generate collections from schema

- Fix OemAccountService to OpenBMCAccountService

This schema as-written broke a bunch of Redfish specification rules for
schemas, naming and other things.  This commit starts to fix them.

It can be used as a starting point for fixing the existing schemas.
This working through testing.

Tested:
   Compilation passed
   Redfish validator passed

The assoicated upstream commits:

https://gerrit.openbmc.org/c/openbmc/bmcweb/+/60757
https://gerrit.openbmc.org/c/openbmc/bmcweb/+/61064
https://gerrit.openbmc.org/c/openbmc/bmcweb/+/60223
https://gerrit.openbmc.org/c/openbmc/bmcweb/+/57441
https://gerrit.openbmc.org/c/openbmc/bmcweb/+/55215

Signed-off-by: Myung Bae <myungbae@us.ibm.com>