Update schemas to 2022.2
For defect SW551681 need LocationIndicatorActive on FabricAdapter.

This is part of the new FabricAdapter schema.

This change is roughly picking up:
https://gerrit.openbmc.org/c/openbmc/bmcweb/+/57360/
https://gerrit.openbmc.org/c/openbmc/bmcweb/+/57415/

Upstream has additional changes that don't need picked up:
https://github.com/openbmc/bmcweb/commits/master/redfish-core/include/schemas.hpp
https://github.com/openbmc/bmcweb/commits/master/scripts/update_schemas.py

Upstream also has an allow list.
Easier to just change 1 line and rerun the script.
The only line manually changed is
https://github.com/ibm-openbmc/bmcweb/blob/1020/scripts/update_schemas.py#L12

Tested: Built an image and ran the validator.

Signed-off-by: Gunnar Mills <gmills@us.ibm.com>