Sort schema list alphabetically (#446)
New versions of the DMTF schema pack seem to make this list parse in
less-than-ideal order.  Previously, the schema pack zip file had an
order that was at least semi-alphabetical, but 2022.2 seems to cause
that to change.

This commit makes the update_schemas.py script explicitly alphabetize
the list.  This changes the order of two keys that have common prefixes,
but leaves the rest of the order the same.  Both are arguably correct.

This commit also exposes that this script seemed to have lots of
problems in its version number detection logic.  To make all this work
properly, a fairly invasive change is required.

Tested: Running the script on 2021.4 produces the same result as
previously.  See later commits for 2022.2 where result is now correct.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: Ia2365daee1d91de142652161bf752600f3109115

Signed-off-by: Ed Tanous <edtanous@google.com>
Co-authored-by: Ed Tanous <edtanous@google.com>