Set Functional property after battery CM (#476)
When bmcweb is told the RTC battery is replaced after a CM operation,
set the Functional property on its assembly object path back to true
before restarting the adcsensor daemon.

This does two things:
1) Causes the fault LED to turn off
2) Sets the Redfish representation back to healthy so the web UI doesn't
   show it has unhealthy it after the CM operation.  Of course if
   adcsensor detects an error again it would go back to unhealthy.

Since this is handling the fault LED, the previous code that explicitly
turned off the fault LED group was removed.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>

Signed-off-by: Matt Spinler <spinler@us.ibm.com>