Fix python packaging problem (#482)
Python packaging seems to have changed to be much more stringent about
what version strings is accepts.  It's not clear why this was done, but
we need a fix to unbreak CI.

This commit adds more parsing code into the script to parse out version
numbers from  the form of LogEntry.v1_2_3 to split them up into a custom
class, and sort manually.  Despite dropping the dependency, this ends up
dropping a net 7 lines of code compared to using the library, so it
seems worth it, especially given the breaking API change.

One seemingly unrelated change was that we were adding schema folders to
the sort lists, which don't sort well because they have no version
numbers or extensions.  This fixes to only add individual files to the
list of things to sort.

Tested:
Code no longer crashes.
Running update_schemas.py generates the same result (with one minor
same-prefix sort order change).

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: I1a792ffb4ad7bc71637ab1aa0648a245949aeeee

Signed-off-by: Ed Tanous <edtanous@google.com>
Co-authored-by: Ed Tanous <edtanous@google.com>