Avoid having stale entries in the lock table in BMC
In the current implementation, the lock records are
continuously stored and persisted. When the session
is removed from HMC, or system abruptly closed or
rebooted, this data is left over in BMC.

This data is causing the lock conflict reports as
mentioned in the following defect:
https://w3.rchland.ibm.com/projects/bestquest/?verb=view&id=SW530642&table=defect

One of the conflicts would be,
say HMC1 takes the lock (LockAll), and the connection is
abruptly disconnected from the bmc, the lock table still
contains the details of this HMC1 and its lock.

Now, HMC2 wants to acquire a lock. But, since the lock
table already has data showing that HMC1 has the locks
(LockAll - although it's disconnected), HMC2 will not be
able to acquire lock.

In order to avoid this, this pr contains changes which
deletes the lock file whenever bmcweb restarts.

Thus, the stale lock entries will always be deleted off from
the bmc and hence resolving the conflicts

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>