Populate BMC state based on state manager property (#520)
The "Status" of the BMC has been historically fairly simple. If the BMC
is at the point where it can respond to the Redfish call then it's "OK"
and it's "On". Its "Status" could be derived based on the startup
systemd target, multi-user.target, as either "Starting" or "Enabled".

However, with the introduction of new BMC states it is no longer this
simple. The following design introduced a new "Quiesced" state for the
BMC[1]

There has was also a new BMC state introduced, "UpdateInProgress", which
indicates the BMC is in the middle of a firmware update.

These new BMC states require additional logic when responding to Redfish
queries for the BMC "Status".

When a BMC is in a "Quiesced" state it is an indication that a critical
BMC service has entered the failed state. It is important for BMC
clients to be aware of this state to take the needed steps to service
the system. For example, on IBM systems, this will trigger a call home
event to IBM service. The undefined aspect of this state is what makes
it so critical to get serviced.

bmcweb has a precedent of abstracting systemd states behind services
like x86-power-control/phosphor-state-manager. It can see it in
functions like getHostState() and getChassisState(). bmcweb could
contain all of the logic to determine if the BMC is being updated or has
a failed critical service but that:
  a) Adds a lot more code to bmcweb and
  b) Goes away from our design of allowing other repositories to
     abstract away the complexity of these states.

Tested:
- GET /redfish/v1/Managers/bmc for different BMC states
- Redfish Validator passed

[1]: https://github.com/openbmc/docs/blob/master/designs/bmc-service-failure-debug-and-recovery.md

Change-Id: I8ceafb7036f41678516ebda3dce5200221bba947

Signed-off-by: Ravi Teja <raviteja28031990@gmail.com>
Co-authored-by: Ravi Teja <raviteja28031990@gmail.com>
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>