Redfish OEM Update Command for Concurrent Update (#141) (#518)
This commit adds a new update command to be used by
the management console to perform concurrent updates.
It works like the current update command, but with
added support for the OnReset apply time setting.

Tested: Ensured that redfish command updates as usual
and updates still occurred while the ApplyTime was in
the OnReset state.

Co-authored-by: zamiseck <33493107+zamiseck@users.noreply.github.com>
Signed-off-by: Myung Bae <myungbae@us.ibm.com>