Remove InternalErrors on usbCodeUpdate GET
usbCodeUpdate isn't a crictical service, if it fails, just leave its
properties off the response. Have a defect where on the 45th code
update, usbCodeUpdate didn't respond (might have still been coming up)
and that caused BMC to return an internalError here. This internalError
led to the HMC putting the BMC in No Connection. Yeah, this seems kind
of fragile.. But this service isn't important so don't let it impact the
HMC.

Tested: None.

Signed-off-by: Gunnar Mills <gmills@us.ibm.com>