Move to  ServiceProviderNotify
This is a cherry-pick of
https://gerrit.openbmc.org/c/openbmc/bmcweb/+/46886 with heavy
modification for downstream. The heavy modification was needed because
we have this separate CElog and we already had this bool
ServiceProviderNotify support.

This is needed because we picked up 3c7b858a in a recent rebase.
https://github.com/ibm-openbmc/phosphor-dbus-interfaces/commit/3c7b858a13aa5fde934d6d789c9a6027490f78dc

And the phosphor-logging change as well.
https://github.ibm.com/openbmc/openbmc/commit/9540c8a68ef3e177cb3cf3d87304e9a7c96178b1#diff-2f4a5d3a11e1bd59bb17797d11c691e415dd4dfd4279facd279d840b6617d2ca

Routes like /redfish/v1/Systems/system/LogServices/CELog were getting an
internalError.

The journal:
eb 01 17:38:16 ever6bmc bmcweb[1808]: (2023-02-01 17:38:16) [CRITICAL
"error_messages.cpp":259] Internal Error
../git/redfish-core/lib/log_services.hpp(2051:60)
`redfish::requestRoutesDBusEventLogEntryCollection(App&)::<lambda(const
crow::Request&, const std::shared_ptr<bmcweb::AsyncResp>&, const
std::string&)>::<lambda(boost::system::error_code, const
dbus::utility::ManagedObjectType&)>`:

The original commit msg:

Add ServiceProviderNotify property that can be used to indicate that
this event log should be notified to the service provider, if that is
implemented on the system.

ServiceProviderNotify is an enum.
Supported options are - "NotSupported", "Notify" and "Inhibit".

The dbus interface and all implementations have been moved to an enum
with the default value of NotSupported. When the value is NotSupported,
this property is left off Redfish.

For more details refer to
https://gerrit.openbmc.org/c/openbmc/phosphor-dbus-interfaces/+/47683

Tested: Manually tested on the system, Run Redfish validator. Found new
error.

Change-Id: I774bfec157481ccc9b4966bf5e8cc8f7d9a06fd0
Signed-off-by: Lakshmi Yadlapati <lakshmiy@us.ibm.com>
Signed-off-by: Abhishek Patel <Abhishek.Patel@ibm.com>
Signed-off-by: Gunnar Mills <gmills@us.ibm.com>