Check PowerSupply availability prop for State/Hlth
This commit makes use of the Available property on the power supply
D-Bus object when determining the power supply state and health in the
PowerSupply response.  The power supply monitor code from phosphor-power
will set the property to false if it can determine that the power supply
isn't able to provide power due to certain faults.

If the PS isn't available, then the Health property will be set to
Critical, and the State will be set to UnavailableOffline (assuming it
shouldn't be Absent instead).

This is necessary because on IBM systems the Functional property
determines the fault LED state as well as the health value, and in
certain cases the fault LED needs to stay off even though the PS health
is not OK.  A specific example of this is when the PS cord is unplugged:
no fault LED is desired but it is desired for the Redfish output to show
that there is an issue with the PS.

If the Available property isn't present on D-Bus, it acts as if it had a
value of true and behaves the same as it does today.

Tested:
Available = false:
    "Health": "Critical",
    "State": "UnavailableOffline"

Available = true, Functional = false:
    "Health": "Critical",
    "State": "Enabled"

PS missing:
    "Health": "Critical",
    "State": "Absent"

Everything OK:
    "Health": "OK",
    "State": "Enabled"

No Available property on D-Bus:
    "Health": "OK",
    "State": "Enabled"

Change-Id: I1a3194bf3a6ca3936954b31439070dcc2f343411

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Signed-off-by: Myung Bae <myungbae@us.ibm.com>