auditEvent(): Prevent overflow of local buffer
If event being audited is larger than will fit in the local buffer,
then the event will be audited but the message will be truncated.
An error message is also sent. If many of these errors are
encountered the local buffer size should be increased or
malloc() should be used.

Tested: I built a version with a much smaller buffer size and
	tested boundaries against the size being used.