Read patch data earlier during an event log patch (#601)
This fixes MalformedJSON errors when trying to patch the Resolved or
ManagementSystemAck properties in an EventLog or CELog.

The first 8 bytes of the request JSON data in the Request structure was
corrupt if it was read after an async D-Bus method call.  Testing showed
that up until the method call was made, it was fine.

This commit changes the code to access that patch data earlier, before
any D-Bus calls are made.  This makes it similar to how it was done for
the Resolved property in the upstream code.

The code had been refactored when the CELog entries were added
downstream to give more commonality, and that is where the change had
been made to read the patch JSON data after reading the Hidden property
on D-Bus.  It did work fine then though, and it also even worked when
the commits were first pulled into 1050.

I'm not sure what changed to make it stop working.  Maybe it was never
valid to access the Request data after code goes back to the event loop
again?

Change-Id: I4872558d42120321aae11f2eddb949e0a14a1886

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I1a0e2887cb1c8e75c8cdd560aed4fdf5911bce99