Remove the default assembly Health
Have a defect where assemblies for the I/O expansion chassis's
assemblies don't have an operationalStatus because we don't know the
health of the FRU but we still have a Redfish Health for these FRUs.
Upstream we have been default Health and State, the thinking is "Hey I
reported this inventory item, mark it as good by default.". Upstream
maybe we should do something different.

Signed-off-by: Gunnar Mills <gmills@us.ibm.com>