Add Support for privilege check in handleUpgrade
This commit enables privilege check for user(s) in case of upgraded
connections.
Currently users with no privileges will also be able to access
Websockets connections (Ex: KVM).

The privilege check was already in place for normal connections (i.e.
router->handle()). This commit lifts off the privilege check code and
moves it into a common function (validatePrivilege()), which can be used
both by handle() and handleUpgrade() and register required callback to
be called.

Also, the const qualifier for Request in the handleUpgrade() function's
signature is removed to enable setting "isConfigureSelf" field of
request. The signature of handleUpgrade() is made identical to handle()

Unfortunately, this does not work for consoles - host/hyp consoles
because the original privilege checking is not done properly in handleUpgrade().

So, at this time, it will keep the same way as it was done 1030
to check "service" for hyp console.

Tested:
 - Redfish validator passed.
 - websocket_test.py
 - Admin and Operator users are able to access console on WebUI
 - Readonly User was unable to access console on WebUI

Example test method of 'console' on webui:
 - Use web-browser, and open 'inspect' sub-page tab
 - Under 'admin',  access & switch between
   https://${bmc-ip}:18080/#/operations/service-login
   https://${bmc-ip}:18080/#/operations/host-console

Signed-off-by: P Dheeraj Srujan Kumar <p.dheeraj.srujan.kumar@intel.com>
Change-Id: I6f743c27e7e6077f1c6c56e6958922027e4404e8

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: Ib6d99726a64326b7c8bad15bc9d4ca774ab6256d

Signed-off-by: Myung Bae <myungbae@us.ibm.com>