Fix and Refactor PCIeDevice handling
There are a few issues in the current 1050 pcie.hpp.

First, it uses a global variable to keep all PCIeDevices,
and it is constructed and populated only once.
Any consecutive PCIeDevice interactions will use it.
As the result, if a new device is added/discovered later,
it will not be found via bmcweb.

Even after the fix of this gobal variable,
the code invokes the unnecessary and excessive calls of GetSubTree.

For example,

```
void getPCIeDevices()
{
  GetSubTree(pcieInterace)
    ...
    getAllDProperties()
        addLinkStatusToPcieDevice()
           { This again go thru GetSubTree() to find pcieSlot <==
             findPcieSlotServiceMap() to get pcieSlot info }
}
```

This can be optimized as

```
void getPCIeDevices()
{
     GetSubTreePaths() to find a device & pcieSlot name
     GetObject => serviceName
     GetAllObject for its properties
     findPcieSlotServiceMap() for pcieSlot related
}
```

Similar optimization can be done in a few other places.

Tested:
 - GUI
 - Redfish validator
 - curl on PCIeDevices & its elements outputs between the existing bmcweb and new code.

Signed-off-by: Myung Bae <myungbae@us.ibm.com>