Util api to fetch pretty name for assemblies
This commit implements change to use util api to fetch prettyName
property for any given assembly and publish it as "Name" in
assembly schema.

In case prettyName is blank for that FRU, then last part of the
object path will be used to populate assembly name.

Validator executed with no new errors.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>