Audit Log: Exclude body data for all IBM management console events
In analyzing the DVT dump of the audit log data I found additional IBM
management console events which contain binary data. Additionally other
IBM management console events contain large body data which overflows
the audit log buffer causing the data to be truncated for the logging.

The body of these events are not providing useful information for the
logging so the body will be excluded for all of these events. (I.e.
URI starting with "/ibm/v1".

The event action will still be logged including the user who initiated the
event.

Signed-off-by: Janet Adkins <janeta@us.ibm.com>