PowerSubsystem: Remove automated test case
Remove the automated gtest test case for the PowerSubsystem.

A previous downstream commit added Allocation properties to the
PowerSubsystem output.

This enhancement required D-Bus calls.  Those calls fail with a SIGSEGV
in the gtest environment.  The commonly used global pointer variable
crow::connections::systemBus is set to nullptr in the gtest environment.

Change-Id: I8aaa9abefc201ceba440dbf8ecff5d5670e27f0b
Signed-off-by: Shawn McCarney <shawnmm@us.ibm.com>