Fix ResourceNotFound on manual guard of Dimm
Issue:
Even after successfully guarding the Dimm, we see
ResourceNotFound Extended Error info

Fix:
When we try to enable or disble a Dimm using redfish, it tries to set the
DimmData for the locationIndicatorActive everytime. This is to be called only
when locationIndicatorActive is provided similar to how it is done for 1030.
So updated the code to match 1030.

Tested and works as expected.