Fix system dump delete functionality
Currently, when a user does a DELETE operation on
any of the system dumps, "Resource Not Found" error is being
sent back by bmcweb.

This is because, all the system dump entry IDs will be of the format:
<dumpType>_<ID>. For example, "Resource_1"/"System_1". Currently, no
logic is present to get dump id from the above string to make
dbus call and the call fails. This issue is resolved in this PR.

Change-Id: Ifde5a532bcc90e6c5c150a704c8e045b8befbf48
Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>