log_services: Fix isContentTypeAllowed checks
The commit https://gerrit.openbmc.org/c/openbmc/bmcweb/+/56694
inadvertently changed the polarity on the if checks using the newly
introduced isContentTypeAllowed function which caused 'Bad Request' to
be returned when the content type was allowed.

Tested:
Getting the EventLog and PostCodes attachment would return the data
instead of 'Bad Request'.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: Iafcdeaba1a0723326347bb2a832b53bbf0aab230