Don't internal error if no slot location code (#683)
https://github.com/ibm-openbmc/bmcweb/pull/650 added PcieSlots Location
Code but it throws an internal error when there is no location code
interface. Location code isn't required and logical slots don't have a
location code so just leave off location code if the interface isn't
there. This is what we had for 1030.

Archana was seeing "Base.1.13.0.InternalError", at
https://${bmc}/redfish/v1/Chassis/chassis15364/PCIeSlots

Tested: Loaded onto the system and didn't see the error anymore.

Signed-off-by: Gunnar Mills <gmills@us.ibm.com>