Audit Log: Correct logging of success login case.
The audit logging of a success login case was happening even if the
login failed. This was because of the placement of the callout to do the
logging. Moved the callout so that the logging only happens after a
pamAuthenticateUser() call. Success or fail reported will depend on
result of the authentication.