Audit Log: Open and save fd for auditing future events
Improve performance of auditLog by opening the audit log file the first
time we try to write an entry. Then reuse the same fd for all other events.

Note: Closing the fd is not necessary as the fd will be closed
automatically when bmcweb exits. A utility function auditClose() is added
to provide a method if wanted to close the auditing connection for other
reasons.

Signed-off-by: Janet Adkins <janeta@us.ibm.com>