Support PATCH on IPv6StaticDefaultGateways (#696)
* Support PATCH on IPv6StaticDefaultGateways

Currently there is no support to setting up the Static Default IPv6
gateway via redfish.

This commit adds IPv6StaticDefaultGateways parameter to the ethernet
interface, on which user can send PATCH request and setup the Static
IPv6 gateway for the interface.

Backend does not support the DefaultGateway6 as array. So this commit
ignores the PrefixLength and just sets the Address on to the dbus
property and corresponding network file.

Tested by:
 PATCH https://${bmc}/redfish/v1/Managers/bmc/EthernetInterfaces/<id>
 -d '{"IPv6StaticDefaultGateways": [{"Address": "2002:903:15F:325:9:3:29:1",
 "PrefixLength": 24}]}'

 PATCH https://${bmc}/redfish/v1/Managers/bmc/EthernetInterfaces/<id>
 -d '{"IPv6StaticDefaultGateways": [{"Address": "2002:903:15F:325:9:3:29:1"}]}'

 PATCH https://${bmc}/redfish/v1/Managers/bmc/EthernetInterfaces/<id>
 -d '{"IPv6StaticDefaultGateways": [{"PrefixLength": 24}]}'
 --> this will return PropertyMissing error

Change-Id: If6aaa6981a9272a733594f0ee313873a09f67758
Signed-off-by: Sunitha Harish <sunithaharish04@gmail.com>

* Support PATCH on IPv6StaticDefaultGateways

Currently there is no support to setting up the Static Default IPv6
gateway via redfish.

This commit adds IPv6StaticDefaultGateways parameter to the ethernet
interface, on which user can send PATCH request and setup the Static
IPv6 gateway for the interface.

Backend does not support the DefaultGateway6 as array. So this commit
ignores the PrefixLength and just sets the Address on to the dbus
property and corresponding network file.

Tested by:
 PATCH https://${bmc}/redfish/v1/Managers/bmc/EthernetInterfaces/<id>
 -d '{"IPv6StaticDefaultGateways": [{"Address": "2002:903:15F:325:9:3:29:1",
 "PrefixLength": 24}]}'

 PATCH https://${bmc}/redfish/v1/Managers/bmc/EthernetInterfaces/<id>
 -d '{"IPv6StaticDefaultGateways": [{"Address": "2002:903:15F:325:9:3:29:1"}]}'

 PATCH https://${bmc}/redfish/v1/Managers/bmc/EthernetInterfaces/<id>
 -d '{"IPv6StaticDefaultGateways": [{"PrefixLength": 24}]}'
 --> this will return PropertyMissing error

Change-Id: If6aaa6981a9272a733594f0ee313873a09f67758
Signed-off-by: Sunitha Harish <sunithaharish04@gmail.com>

---------

Signed-off-by: Sunitha Harish <sunithaharish04@gmail.com>
Co-authored-by: Sunitha Harish <sunithaharish04@gmail.com>