chassis: consistently log error on internal fails
Debugging internalError responses from bmcweb has been a consistent
pain point for us. It does help to at least have the boost error code
logged to the journal. Ensure the error code is logged to the journal
consistently within the chassis file.

If this commit makes sense to the maintainers then I will work on making
this consistent in other files at a later date.

Tested:
- Verified it compiled and one of the error paths traced the appropriate
  boost error code.

Change-Id: I76b7644bc18e76fff69595fb5f1bc23d257563e9
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>