Audit Log: Handle breaking of unexpected connection with auditd
If something breaks the connection for the auditfd socket, then events
will fail to be audited. Even if the socket connection could be
established again, there would not be another attempt to establish the
connection.

To handle the error I've altered the semantics of auditOpen(). It will
no longer prevent another attempt to connect to the auditd subsystem
after the first connection attempt.

New auditReopen() function is added to allow a retry on a failed audit
event. auditClose() can still be used to block further auditing
attempts.

Unit test added for auditReopen().

Signed-off-by: Janet Adkins <janeta@us.ibm.com>