Fix PCIeSlots Redfish validator Error
The latest Redfish-Validator reports the following errors.

```
26 err.Collection(Processor.Processor) errors in /redfish/v1/Chassis/chassis/PCIeSlots
16 failInvalidArray errors in /redfish/v1/Chassis/chassis/PCIeSlots
26 failProp errors in /redfish/v1/Chassis/chassis/PCIeSlots

ERROR - Property AssociatedAssembly should not be a List
ERROR - This object's type AssociatedAssembly should be a Collection, but it's of type Resource.OemObject...
ERROR - AssociatedAssembly: Value of property is an array but is not a Collection
ERROR - AssociatedAssembly: Value of property is an array but is not a Collection
...
WARNING - Couldn't get schema for object (?), skipping OemObject IBM : 'IBM'
WARNING - Couldn't get schema for object (?), skipping OemObject IBM : 'IBM'
```

It is caused because `AssociatedAssembly` is incorrectly showing as an
array.

```
$ curl -k -X GET https://${bmc}/redfish/v1/Chassis/chassis/PCIeSlots
  ...
            "AssociatedAssembly": [
              {
                "@odata.id": "/redfish/v1/Chassis/chassis/Assembly#/Assemblies/17"
              }
            ]

==>
            "AssociatedAssembly": {
              "@odata.id": "/redfish/v1/Chassis/chassis/Assembly%23/Assemblies/17"
            }
```

Tested:
- Redfish validator runs without the reported errors
- Redfish validator reduces warnings

Signed-off-by: Myung Bae <myungbae@us.ibm.com>