Push Style Event: Cleanup at maxRequestQueueSize
HMC subscribes with keep_alive and RetryForever option.
At some unusual circumstances, HMC abruptly closes the socket
and does not send error response, BMC keeps waiting for the
response and keeps adding the events to send queue.
This ultimately results in not sending the events to HMC and hits
max queue size.

This change is to clear the queue at this point and close the
connection and do related cleanups. The next event will be sent
good by creating new socket and connections.

Tested by:
 Tested event sanity by connecting a HMC and performing 10
 iterations of PowerOn and PoweOff.
 Could not recreate the original issue with this patch

Signed-off-by: Sunitha Harish <sunithaharish04@gmail.com>