Fix the removal of unrelated queued http request
During RetryForever support implementation, the front queued item is
removed if the queud becomes full, regardless whether it is related to
RetryForever option or not.

This logic may cause the side-effect.
For example, theoretically, if the queued item is still being
handled and left in the async-call schedulers, the removed object may
be referenced when the scheduler pops.

This change is to preserve the queued items to be active as long as
they are still connected.  If the newer requests cause the queue
pool overflowed, they will be rejected (rather than the queue item
is to removed).

Signed-off-by: Myung Bae <myungbae@us.ibm.com>