1060: Move to boost 1.83 & clang17 (#781)
* Update to boost 1.83.0

In boost 1.83.0, the boost::url maintainers deprecated the header only
usage of the library without warning.  A discussion with the
maintainers[1] made it clear that they removed the abiliy on purpose,
and they're not going to add it back or add a deprecation strategy (they
did say they would update the documentation to actually match the
intent), and that from here on in we should be using the cmake boost
project to pull in the non-header-only boost libraries we use (which at
this point is ONLY boost url).

This commit updates to remove the usage of boost::urls::result typedef,
which was deprecated in this release (which causes a compile error) and
moves it to boost::system::result.

In addition, it updates our meson files to pull in the boost project as
a cmake dependency.

[1] https://cpplang.slack.com/archives/C01JR6C9C4U/p1696441238739129

Tested: Not yet.

Change-Id: Ia7adfc0348588915440687c3ab83a1de3e6b845a
Signed-off-by: Ed Tanous <edtanous@google.com>
Signed-off-by: Myung Bae <myungbae@us.ibm.com>

* Fix local compile

The update to boost 1.83.0 was breaking for our build process if boost
1.83.0 was not already installed.  Update our meson file to correctly
pull in all of the required boost libraries.

Tested:
I was able to locally build bmcweb without having previously installed
any boost libraries.  All unit tests also passed.

meson buildlocal && ninja -C buildlocal test

Signed-off-by: Carson Labrado <clabrado@google.com>
Change-Id: I1d00ef561fed7e3ba799969a112ee58b6578ce32

* Fix boost build locally

Change I1d00ef561fed7e3ba799969a112ee58b6578ce32 fixed local
compilation of bmcweb, but it is causing failures when attempting to
SRCREV bump.  This should fix the issues during SRCREV bump while still
allowing local builds to succeed.

Tested (Carson):
I was able to locally build bmcweb without having previously installed
any boost libraries.  All unit tests also passed.

meson buildlocal && ninja -C buildlocal test

Change-Id: I8cbb764dc564b84b4adb06ba80cb87b2a2c4701f
Signed-off-by: Ed Tanous <edtanous@google.com>
Signed-off-by: Carson Labrado <clabrado@google.com>

* Don't rely on implicit operator comparison

Json has a bunch of implicit overloaded operators for comparing
vectors/arrays/maps with their json counterparts.  Unfortunately, when
used in unit tests, these cause warnings in clang, so update the code to
use the ElementsAre check from gmock.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: I658557cb59d568fd50cf6f3bef73d6f90b5c56cf

* hex_units: Fix Werror=conversion

Convert all types to uint8_t to not hit the conversion warning.

Change-Id: Ia535ca0a2f4045cbde06a2f8f8eaad9570a0f4a5
Signed-off-by: Willy Tu <wltu@google.com>
Signed-off-by: Ed Tanous <edtanous@google.com>

* Aggregation: Fix cppcheck errors

Corrections style complaints in the aggregator code.

Tested:
Jenkins output did not show any style complaints

Signed-off-by: Carson Labrado <clabrado@google.com>
Change-Id: I87426fcf2a48448a62152e0ad4a6c3aa54a7fc45

* 1060: clang-format updates for clang17

This is from:
 https://gerrit.openbmc.org/c/openbmc/bmcweb/+/67270/3/.clang-format

Can't cherry pick the commit due to conflicts so just pull in the
.clang-format file and run against downstream code.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>

---------

Signed-off-by: Ed Tanous <edtanous@google.com>
Signed-off-by: Myung Bae <myungbae@us.ibm.com>
Signed-off-by: Carson Labrado <clabrado@google.com>
Signed-off-by: Willy Tu <wltu@google.com>
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Co-authored-by: Ed Tanous <edtanous@google.com>
Co-authored-by: Carson Labrado <clabrado@google.com>
Co-authored-by: Willy Tu <wltu@google.com>