util: Add pretty name for resources (#918)
Created a helper function to fetch pretty name and added for some
resources.

- Chassis
- FirmwareInventory
- Storage
- StorageController
- System
- Processor
- Fan
- Fabric Adaptor
- Memory
- Assembly

Implementation Details:
The helper function makes a call to ObjectMapper to get the services
that support the object path with the interfaces we want. It expects
only one service is exporting the object path.

Then it will try to get the `PrettyName` property. If it is found and is
valid, it will set the `Name` to the value. If not, it will use the
default value that is set before the function call.

This will affect all systems and replace the Name with the PrettyName if
it exist within the resource's object path.

Tested:
Passed Redfish Validator

The output names release the existing `Name` property only if the
PrettyName property exists. If not, it will use the default value.

Output with PrettyName
```
$ curl  -u root:0penBmc -X GET http://${bmc}/redfish/v1/Chassis/chassis0
{
  "@odata.id": "/redfish/v1/Chassis/chassis0",
  "@odata.type": "#Chassis.v1_14_0.Chassis",
  "ChassisType": "RackMount",
  "Id": "chassis0",
  "Links": {
    "ComputerSystems": [
      {
        "@odata.id": "/redfish/v1/Systems/system"
      }
    ],
    "ManagedBy": [
      {
        "@odata.id": "/redfish/v1/Managers/bmc"
      }
    ]
  },
  "Name": "Test Chassis",
  ...
}
```

Without PrettyName
```
$ curl  -u root:0penBmc -X GET http://${bmc}/redfish/v1/Chassis/chassis0
{
  "@odata.id": "/redfish/v1/Chassis/chassis0",
  "@odata.type": "#Chassis.v1_14_0.Chassis",
  "ChassisType": "RackMount",
  "Id": "chassis0",
  "Links": {
    "ComputerSystems": [
      {
        "@odata.id": "/redfish/v1/Systems/system"
      }
    ],
    "ManagedBy": [
      {
        "@odata.id": "/redfish/v1/Managers/bmc"
      }
    ]
  },
  "Name": "chassis0",
  ...
}
```

Squashed following commits:
Logging  invalid service name for debug
Util api to fetch pretty name for assemblies
Dimm: Look at Pretty Name (#973) (#1111)
Add pretty name for fabric adapters (#869)

The method getPrettyName() may be called more than one time.
This will generate journal logging messages for the required Service name
if there are unexpected service names.

In case prettyName is blank for that FRU, then last part of the
object path will be used to populate assembly name.

Signed-off-by: Willy Tu <wltu@google.com>
Co-authored-by: Willy Tu <wltu@google.com>
Co-authored-by: Sunny Srivastava <sunnsr25@in.ibm.com>
Co-authored-by: Gunnar Mills <gmills@us.ibm.com>