Fix 404 on Hypervisor Reset
/redfish/v1/Systems/hypervisor/Actions/ComputerSystem.Reset was
returning a 404. On closer inspection, it was due to hitting the System
.Reset Action and not making it to the Hypervisor specific action. [1].

Remove the Hypervisor specific action and call the method from the
System .Reset Action. This is how we do the handleHypervisorSystemGet
already. [2]

Fixes defects 637587 and 632508.

[1]: https://github.com/ibm-openbmc/bmcweb/blob/1110/redfish-core/lib/systems.hpp#L3335
[2]: https://github.com/ibm-openbmc/bmcweb/blob/1110/redfish-core/lib/systems.hpp#L2906C9-L2906C34

Tested: Loaded this on a system and was able to "Continue to OS
running".

Signed-off-by: Gunnar Mills <gmills@us.ibm.com>