Generate secret key: Remove global mfa check
When a user generates secret key, currently there is a check to see if
the global mfa is enabled. And only if it is enabled, secret key
generation is allowed for the user.
This commit removes the check, and enables users to generate secret key
even if the global mfa is disabled.

Tested By:

'''
POST https://${bmc}/redfish/v1/AccountService/Accounts/<>/Actions/ManagerAccount.GenerateSecretKey
'''

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>